tools/can_packet_parser/requirements.txt
digi-xbee==1.4.0
pyserial==3.5
python-can==4.0.0
pyyaml==6.0
matplotlib==3.5.2